create separated services for courses and enrollments in order to simplify the logic in professor service
