+++ src/main/java/bussiness_layer/services/IStudentService.java
@@ -1,15 +1,5 @@
 package bussiness_layer.services;
 
-import bussiness_layer.dto.EnrollmentDto;
-
-import java.util.List;
-
 public interface IStudentService {
-    /**
-     * Get all the enrollments that belong to a student.
-     *
-     * @param studUsername - the username to identify the student
-     * @return - the enrollments for the  given student
-     */
-    List<EnrollmentDto> getEnrollments(String studUsername);
+
 }
