+++ src/main/java/bussiness_layer/services/impl/AuthenticationService.java
@@ -1,4 +1,4 @@
-package bussiness_layer.services;
+package bussiness_layer.services.impl;
 
 import data_layer.domain.User;
 import data_layer.repositories.IUserRepository;
