+++ src/main/java/data_layer/domain/User.java
@@ -1,15 +1,22 @@
 package data_layer.domain;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.persistence.Table;
+import javax.persistence.UniqueConstraint;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 @Data
 @Entity(name = "Users")
-@Table(uniqueConstraints = @UniqueConstraint(columnNames = "username"))
+@Table(uniqueConstraints = {@UniqueConstraint(columnNames = {"username"})})
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 @NoArgsConstructor
 public abstract class User {
