+++ src/main/java/web_layer/controllers/ProfessorMvcController.java
@@ -4,6 +4,7 @@ import bussiness_layer.dto.EnrollmentDto;
 import bussiness_layer.dto.GroupDto;
 import bussiness_layer.dto.ProfessorCourseDto;
 import bussiness_layer.dto.ProfessorRightDto;
+import bussiness_layer.services.IEnrollmentService;
 import bussiness_layer.services.IProfessorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,11 +22,16 @@ import java.util.List;
 public class ProfessorMvcController {
 
     @Autowired
-    private IProfessorService service;
+    private final IProfessorService professorService;
 
-    public ProfessorMvcController(IProfessorService service) {
-        this.service = service;
+    @Autowired
+    private final IEnrollmentService enrollmentService;
+
+    public ProfessorMvcController(IProfessorService professorService, IEnrollmentService enrollmentService) {
+        this.professorService = professorService;
+        this.enrollmentService = enrollmentService;
     }
+
 
     @GetMapping("/")
     public String getIndexPage() {
@@ -34,7 +40,7 @@ public class ProfessorMvcController {
 
     @GetMapping("/dashboard")
     public ModelAndView getDashboardPage(Principal crtUser) {
-        List<ProfessorCourseDto> professorCourses = service.getRelatedCourses(crtUser.getName());
+        List<ProfessorCourseDto> professorCourses = professorService.getRelatedCourses(crtUser.getName());
         return new ModelAndView("/professor/dashboard")
                 .addObject("professorCourses", professorCourses)
                 .addObject("viewHelper", ViewHelper.class);
@@ -42,9 +48,9 @@ public class ProfessorMvcController {
 
     @GetMapping("/timeline")
     public ModelAndView getTimelinePage(@RequestParam String course, @RequestParam String group, Principal crtUser) {
-        List<EnrollmentDto> enrollments = service.getEnrollments(crtUser.getName(), course, group);
-        List<ProfessorRightDto> professorRights = service.getProfessorRights(crtUser.getName(), course, group);
-        List<GroupDto> groups = service.getGroups(crtUser.getName(), course);
+        List<EnrollmentDto> enrollments = enrollmentService.getEnrollments(crtUser.getName(), course, group);
+        List<ProfessorRightDto> professorRights = professorService.getProfessorRights(crtUser.getName(), course, group);
+        List<GroupDto> groups = professorService.getGroups(crtUser.getName(), course);
         return new ModelAndView("/professor/timeline")
                 .addObject("enrollments", enrollments)
                 .addObject("rights", professorRights)
