+++ src/main/java/web_layer/controllers/ProfessorRestController.java
@@ -2,6 +2,7 @@ package web_layer.controllers;
 
 import bussiness_layer.dto.CourseDto;
 import bussiness_layer.dto.LessonDto;
+import bussiness_layer.services.ICourseService;
 import bussiness_layer.services.IProfessorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -15,33 +16,37 @@ import java.util.ArrayList;
 public class ProfessorRestController {
 
     @Autowired
-    private final IProfessorService service;
+    private final IProfessorService professorService;
 
-    public ProfessorRestController(IProfessorService service) {
-        this.service = service;
+    @Autowired
+    private final ICourseService courseService;
+
+    public ProfessorRestController(IProfessorService professorService, ICourseService courseService) {
+        this.professorService = professorService;
+        this.courseService = courseService;
     }
 
     @PutMapping("/lessons")
     public ResponseEntity<?> putLessons(@RequestBody ArrayList<LessonDto> lessons, Principal crtUser) {
-        service.updateLessons(crtUser.getName(), lessons);
+        professorService.updateLessons(crtUser.getName(), lessons);
         return ResponseEntity.ok(null);
     }
 
     @GetMapping("/courses")
     public ResponseEntity<?> getRelatedCourses(Principal crtUser) {
-        service.getRelatedCourses(crtUser.getName());
+        professorService.getRelatedCourses(crtUser.getName());
         return ResponseEntity.ok(null);
     }
 
     @PutMapping("/course")
     public ResponseEntity<?> putCourse(@RequestBody CourseDto course, Principal crtUser) {
-        service.updateCourse(crtUser.getName(), course);
+        courseService.updateCourse(crtUser.getName(), course);
         return ResponseEntity.ok(null);
     }
 
     @GetMapping("/course/{courseCode}")
     public CourseDto getCourse(@PathVariable String courseCode, Principal crtUser) {
-        return service.getCourse(crtUser.getName(), courseCode);
+        return courseService.getCourse(crtUser.getName(), courseCode);
     }
 
 }
