+++ src/main/java/web_layer/controllers/StudentMvcController.java
@@ -1,22 +1,23 @@
 package web_layer.controllers;
 
-import bussiness_layer.dto.EnrollmentDto;
-import bussiness_layer.services.IStudentService;
+import java.security.Principal;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-import web_layer.utils.ViewHelper;
 
-import java.security.Principal;
-import java.util.List;
+import bussiness_layer.dto.EnrollmentDto;
+import bussiness_layer.services.IEnrollmentService;
+import web_layer.utils.ViewHelper;
 
 @Controller
 @RequestMapping("/student")
 public class StudentMvcController {
     @Autowired
-    private IStudentService service;
+    private IEnrollmentService enrollmentService;
 
     @GetMapping("/")
     public String getIndexPage() {
@@ -25,9 +26,10 @@ public class StudentMvcController {
 
     @GetMapping("/timeline")
     public ModelAndView getTimelinePage(Principal crtUser) {
-        List<EnrollmentDto> enrollments = service.getEnrollments(crtUser.getName());
+        List<EnrollmentDto> enrollments = enrollmentService.getEnrollments(crtUser.getName());
         return new ModelAndView("/student/timeline")
                 .addObject("enrollments", enrollments)
                 .addObject("viewHelper", ViewHelper.class);
     }
+
 }
