+++ src/test/java/business_layer/BaseIntegrationTest.java
@@ -1,20 +1,44 @@
 package business_layer;
 
-import bd_config.H2TestConfiguration;
-import data_layer.domain.*;
-import data_layer.repositories.*;
-import factory.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.Random;
+
+import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
+
+import bd_config.H2TestConfiguration;
+import data_layer.domain.Course;
+import data_layer.domain.Enrollment;
+import data_layer.domain.Group;
+import data_layer.domain.Lesson;
+import data_layer.domain.LessonTemplate;
+import data_layer.domain.Professor;
+import data_layer.domain.ProfessorRight;
+import data_layer.domain.Student;
+import data_layer.domain.User;
+import data_layer.repositories.ICourseRepository;
+import data_layer.repositories.IEnrollmentRepository;
+import data_layer.repositories.IGroupRepository;
+import data_layer.repositories.ILessonRepository;
+import data_layer.repositories.ILessonTemplateRepository;
+import data_layer.repositories.IProfessorRightRepository;
+import data_layer.repositories.IUserRepository;
+import factory.CourseFactory;
+import factory.GroupFactory;
+import factory.LessonFactory;
+import factory.LessonTemplateFactory;
+import factory.ProfessorFactory;
+import factory.ProfessorRightFactory;
+import factory.StudentFactory;
 import utils.LessonType;
 import utils.RightType;
 import utils.TestConstants;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * Utilitary class for tests.
@@ -45,24 +69,52 @@ public abstract class BaseIntegrationTest {
     @Autowired
     protected ILessonTemplateRepository lessonTemplateRepository;
 
+    @Before
+    public void emptyDatabase() {
+        lessonRepository.deleteAll();
+        lessonRepository.flush();
+        enrollmentRepository.deleteAll();
+        enrollmentRepository.flush();
+        lessonTemplateRepository.deleteAll();
+        lessonTemplateRepository.flush();
+        professorRightRepository.deleteAll();
+        professorRightRepository.flush();
+        courseRepository.deleteAll();
+        courseRepository.flush();
+        userRepository.deleteAll();
+        userRepository.flush();
+        groupRepository.deleteAll();
+        groupRepository.flush();
+    }
 
     protected User createUser(User u) {
         return userRepository.save(u);
     }
 
     protected Enrollment createEnrollment(String profUsername, String courseCode, String groupCode) {
-        return createEnrollment(profUsername, courseCode, groupCode, TestConstants.USERNAME, TestConstants.USERNAME);
+        return createEnrollment(profUsername, courseCode, groupCode, TestConstants.PROF_USERNAME, TestConstants.STUD_USERNAME);
     }
 
     protected Enrollment createEnrollment(String profUsername, String courseCode, String groupCode
             , String courseCoordinatorUsername, String studentUsername) {
-        Professor professor = userRepository.save(ProfessorFactory.generateProfessorBuilder()
+        userRepository.save(ProfessorFactory.generateProfessorBuilder()
                 .username(profUsername)
                 .build());
-        Course course = courseRepository.save(CourseFactory.generateCourseBuilder()
-                .coordinator(userRepository.save(ProfessorFactory.generateProfessorBuilder().username(courseCoordinatorUsername).build()))
+
+        Course course = CourseFactory.generateCourseBuilder()
                 .code(courseCode)
-                .build());
+                .build();
+        Professor professor = null;
+
+        Optional<User> courseCoordinator = userRepository.findByUsername(courseCoordinatorUsername);
+        if (courseCoordinator.isPresent()) {
+            course.setCoordinator((Professor) courseCoordinator.get());
+        } else {
+            professor = userRepository.save(ProfessorFactory.generateProfessorBuilder().username(courseCoordinatorUsername).build());
+            course.setCoordinator(professor);
+        }
+        course = courseRepository.save(course);
+
         Group group = groupRepository.save(GroupFactory.generateGroupBuilder()
                 .code(groupCode)
                 .build());
