+++ src/test/java/utils/TestConstants.java
@@ -7,12 +7,18 @@ public class TestConstants {
     public static final String FIRSTNAME = "firstname";
     public static final String LASTNAME = "lastname";
     public static final String EMAIL = "abcd1234@scs.ubbcluj.ro";
+
     public static final String FATHERS_INITIALS = "A.B.";
-    public static final String PROF_USERNAME = "prof";
     public static final String COURSE_CODE = "MLR0000";
     public static final String COURSE_CODE2 = "MLR0001";
     public static final String GROUP_CODE = "231";
     public static final String GROUP_CODE2 = "232";
+
+    public static final String PROF_USERNAME = "prof";
+    public static final String PROF_USERNAME2 = "prof2";
+    public static final String PROF_COORDINATOR = "prof coordinator";
+
     public static final String STUD_USERNAME = "stud";
+    public static final String STUD_USERNAME2 = "stud2";
 
 }
