+++ src/test/java/web_layer/BaseControllerTest.java
@@ -5,13 +5,17 @@ import org.junit.runner.RunWith;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
+import lombok.extern.slf4j.Slf4j;
+
 @RunWith(SpringRunner.class)
+@Slf4j
 public class BaseControllerTest {
 
     protected MockMvc mockMvc;
 
     @Test
     public void initializerTestMethod() {
+        log.info("test in BaseControllerTest");
     }
 
 }
